+++ JejuLantis/src/main/java/com/khd/jejulantis/admin/cardetail/DAO/CarkindDetailDao.java
@@ -6,17 +6,17 @@ import com.khd.jejulantis.model.CarkindDetail;
 
 public interface CarkindDetailDao {
 	List<CarkindDetail> carkind(String manager_id);
-	void delete(int car_kind_no);
-	List<CarkindDetail> carkind(int car_kind_no);
+	void delete(long car_kind_no);
+	List<CarkindDetail> carkind(long car_kind_no);
 	List<CarkindDetail> carkindI();
 	void insert(CarkindDetail carkind);
-	List<CarkindDetail> carkindSelectBox(int car_kind_no);
-	List<CarkindDetail> insuranceUpdateSelectBox(String manager_id, int car_kind_no);
+	List<CarkindDetail> carkindSelectBox(long car_kind_no);
+	List<CarkindDetail> insuranceUpdateSelectBox(String manager_id, long car_kind_no);
 	void update(CarkindDetail carkind);
 	List<CarkindDetail> insuranceSelectBox(String manager_id);
 	List<CarkindDetail> branchNoSelect(String manager_id);
 	List<CarkindDetail> kingbranchNotSelect(String manager_id);
 	List<CarkindDetail> identyBNo(String manager_id);
 	List<CarkindDetail> listAll();
-	void update(int car_kind_no, double car_kind_set1);
+	void update(long car_kind_no, double car_kind_set1);
 }
