+++ JejuLantis/src/main/java/com/khd/jejulantis/client/review/DAO/CReviewDAO.java
@@ -7,5 +7,5 @@ import com.khd.jejulantis.model.BranchName;
 public interface CReviewDAO {
 	List<BranchName> branchName();
 	List<BranchName> select(String review_sel);
-	List<BranchName> reviewContentAll(int branch_no);
+	List<BranchName> reviewContentAll(long branch_no);
 }
