+++ JejuLantis/src/main/java/com/khd/jejulantis/model/ReviewContent.java
@@ -15,16 +15,16 @@ public class ReviewContent {
 	private Date rent_reserv_end;
 	private String car_image;
 	//관리자페이지 이용후기 상세보기 
-	private int rent_review_no;
-	private int rent_reserv_no;
-	private int driver_line_no;
-	private int rent_payment_no;
-	private int member_no;
-	private int car_no;
-	private int car_info_no;
-	private int car_kind_no;
-	private int branch_no;
-	private int insurance_no;
+	private long rent_review_no;
+	private long rent_reserv_no;
+	private long driver_line_no;
+	private long rent_payment_no;
+	private long member_no;
+	private long car_no;
+	private long car_info_no;
+	private long car_kind_no;
+	private long branch_no;
+	private long insurance_no;
 	private String rent_review_title;
 	private String rent_review_content;
 	private double rent_review_rating_car;
@@ -32,7 +32,7 @@ public class ReviewContent {
 	
 	public ReviewContent() {}
 	//관리자페이지 이용후기 상세보기 
-	public ReviewContent(String member_name, int rent_review_no, Date rent_review_date, String branch_name,
+	public ReviewContent(String member_name, long rent_review_no, Date rent_review_date, String branch_name,
 			String name, Double rent_review_rating_car, String rent_review_content, String car_name,
 			String car_fuel, String car_type, Date rent_reserv_start, Date rent_reserv_end,
 			String car_image) {
@@ -53,8 +53,8 @@ public class ReviewContent {
 		this.rent_reserv_end = rent_reserv_end;
 		this.car_image = car_image;
 	}
-	public ReviewContent(int rent_review_no, int rent_reserv_no, int driver_line_no, int rent_payment_no,
-			int member_no, int car_info_no, int car_kind_no, int branch_no, int insurance_no,
+	public ReviewContent(long rent_review_no, long rent_reserv_no, long driver_line_no, long rent_payment_no,
+			long member_no, long car_info_no, long car_kind_no, long branch_no, long insurance_no,
 			String rent_review_title, String rent_review_content, Double rent_review_rating_car,
 			Date rent_review_date) {
 		super();
@@ -75,8 +75,8 @@ public class ReviewContent {
 	
 	public ReviewContent(String member_name, String branch_name, String name, String car_name, String car_fuel,
 			String car_type, String insurance_name, Date rent_reserv_start, Date rent_reserv_end, String car_image,
-			int rent_review_no, int rent_reserv_no, int driver_line_no, int rent_payment_no, int member_no, int car_no,
-			int car_info_no, int car_kind_no, int branch_no, int insurance_no, String rent_review_title,
+			long rent_review_no, long rent_reserv_no, long driver_line_no, long rent_payment_no, long member_no, long car_no,
+			long car_info_no, long car_kind_no, long branch_no, long insurance_no, String rent_review_title,
 			String rent_review_content, Double rent_review_rating_car, Date rent_review_date) {
 		super();
 		this.member_name = member_name;
@@ -104,16 +104,16 @@ public class ReviewContent {
 		this.rent_review_rating_car = rent_review_rating_car;
 		this.rent_review_date = rent_review_date;
 	}
-	public int getDriver_line_no() {
+	public long getDriver_line_no() {
 		return driver_line_no;
 	}
-	public void setDriver_line_no(int driver_line_no) {
+	public void setDriver_line_no(long driver_line_no) {
 		this.driver_line_no = driver_line_no;
 	}
-	public int getCar_no() {
+	public long getCar_no() {
 		return car_no;
 	}
-	public void setCar_no(int car_no) {
+	public void setCar_no(long car_no) {
 		this.car_no = car_no;
 	}
 	public String getMember_name() {
@@ -176,58 +176,58 @@ public class ReviewContent {
 	public void setCar_image(String car_image) {
 		this.car_image = car_image;
 	}
-	public int getRent_review_no() {
+	public long getRent_review_no() {
 		return rent_review_no;
 	}
-	public void setRent_review_no(int rent_review_no) {
+	public void setRent_review_no(long rent_review_no) {
 		this.rent_review_no = rent_review_no;
 	}
-	public int getRent_reserv_no() {
+	public long getRent_reserv_no() {
 		return rent_reserv_no;
 	}
-	public void setRent_reserv_no(int rent_reserv_no) {
+	public void setRent_reserv_no(long rent_reserv_no) {
 		this.rent_reserv_no = rent_reserv_no;
 	}
-	public int getdriver_line_no() {
+	public long getdriver_line_no() {
 		return driver_line_no;
 	}
-	public void setdriver_line_no(int driver_line_no) {
+	public void setdriver_line_no(long driver_line_no) {
 		this.driver_line_no = driver_line_no;
 	}
-	public int getRent_payment_no() {
+	public long getRent_payment_no() {
 		return rent_payment_no;
 	}
-	public void setRent_payment_no(int rent_payment_no) {
+	public void setRent_payment_no(long rent_payment_no) {
 		this.rent_payment_no = rent_payment_no;
 	}
-	public int getMember_no() {
+	public long getMember_no() {
 		return member_no;
 	}
-	public void setMember_no(int member_no) {
+	public void setMember_no(long member_no) {
 		this.member_no = member_no;
 	}
-	public int getCar_info_no() {
+	public long getCar_info_no() {
 		return car_info_no;
 	}
-	public void setCar_info_no(int car_info_no) {
+	public void setCar_info_no(long car_info_no) {
 		this.car_info_no = car_info_no;
 	}
-	public int getCar_kind_no() {
+	public long getCar_kind_no() {
 		return car_kind_no;
 	}
-	public void setCar_kind_no(int car_kind_no) {
+	public void setCar_kind_no(long car_kind_no) {
 		this.car_kind_no = car_kind_no;
 	}
-	public int getBranch_no() {
+	public long getBranch_no() {
 		return branch_no;
 	}
-	public void setBranch_no(int branch_no) {
+	public void setBranch_no(long branch_no) {
 		this.branch_no = branch_no;
 	}
-	public int getInsurance_no() {
+	public long getInsurance_no() {
 		return insurance_no;
 	}
-	public void setInsurance_no(int insurance_no) {
+	public void setInsurance_no(long insurance_no) {
 		this.insurance_no = insurance_no;
 	}
 	public String getRent_review_title() {
